CM-208: refactored interface file, added empty group converter
