+++ calcrule_social_protection/converters/__init__.py
@@ -1,2 +1,2 @@
-from calcrule_social_protection.converters.beneficiary_to_bill import BeneficiaryToBillConverter
-from calcrule_social_protection.converters.beneficiary_to_bill_item import BeneficiaryToBillItemConverter
+from calcrule_social_protection.converters.beneficiary import BeneficiaryToBillConverter
+from calcrule_social_protection.converters.beneficiary import BeneficiaryToBillItemConverter
