+++ calcrule_social_protection/converters/beneficiary/__init__.py
@@ -1,0 +1,2 @@
+from calcrule_social_protection.converters.beneficiary.beneficiary_to_bill import BeneficiaryToBillConverter
+from calcrule_social_protection.converters.beneficiary.beneficiary_to_bill_item import BeneficiaryToBillItemConverter
